\name{Biomod.Models}
\alias{Biomod.Models}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
Biomod.Models(Model, Ids, Pa.samp, TypeGLM, Test, No.trees, CV.tree, CV.ann, Perc025, Perc05, NbRunEval, Spline, DataSplit, Yweights, Roc, Optimized.Threshold.Roc, Kappa, TSS, KeepPredIndependent, VarImport)
}

\arguments{
  \item{Model}{  }
  \item{Ids}{  }
  \item{PA.samp}{  }
  \item{TypeGLM}{  }
  \item{Test}{  }
  \item{No.trees}{  }
  \item{CV.tree}{  }
  \item{CV.ann}{  }
  \item{Perc025}{  }
  \item{Perc05}{  }
  \item{NbRunEval}{  }
  \item{Spline}{  }
  \item{DataSplit}{  }
  \item{Yweights}{  }
  \item{Roc}{  }
  \item{Optimized.Threshold.Roc}{ }
  \item{Kappa}{  }
  \item{TSS}{  }
  \item{KeepPredIndependent}{  }
  \item{VarImport}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
