\name{CV.nnet}
\alias{CV.nnet}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}


\usage{
CV.nnet(Input, Target, size=c(2,4,6, 8), decay=c(0.001, 0.01, 0.05, 0.1), nbCV=5, W=NULL)
}

\arguments{
  \item{Input}{  }
  \item{Target}{  }
  \item{size}{  }
  \item{decay}{  }
  \item{nbCV}{  }
  \item{W}{  }
}
\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
