\name{CVnnet}
\alias{CVnnet}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}


\usage{
CVnnet(Input, Target, size=c(2,4,6, 8), decay=c(0.001, 0.01, 0.05, 0.1))
}

\arguments{
  \item{formula}{  }
  \item{data}{  }
  \item{truth}{  }
  \item{size}{  }
  \item{lambda}{  }
  \item{nreps}{  }
  \item{nifold}{  }
  \item{\dots}{  }
}
\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
