\name{CVnnet1}
\alias{CVnnet1}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}

\usage{
CVnnet1(formula, data, truth, nreps = 1, ri, ...)
}

\arguments{
  \item{formula}{  }
  \item{data}{  }
  \item{truth}{  }
  \item{nreps}{  }
  \item{ri}{  }
  \item{\dots}{  }
}
  
\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }