\name{Dispersal.Limit}
\alias{Dispersal.Limit}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
Dispersal.Limit(CurrentPred = NULL, FutureProj = NULL, X = NULL, Y = NULL, MaxMigr = NULL)
}

\arguments{
  \item{CurrentPred}{  }
  \item{FutureProj}{  }
  \item{X}{  }
  \item{Y}{  }
  \item{MaxMigr}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
