\name{FilteringTransformation}
\alias{FilteringTransformation}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
FilteringTransformation(ProbData, CutOffdata)
}

\arguments{
  \item{ProbData}{  }
  \item{CutOffdata}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }

