\name{functionkeep}
\alias{functionkeep}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}

\usage{
functionkeep(object, AIC)
}

\arguments{
  \item{object}{  }
  \item{AIC}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
