\name{Future1}
\alias{Future1}
\docType{data}
\title{ Future data for making projections }
\description{ It's just some data, come on
  
}
\usage{data(Future1)}
\format{
  A data frame with 2264 observations on the following 8 variables.
  \describe{
    \item{\code{Id}}{a numeric vector}
    \item{\code{Var1}}{a numeric vector}
    \item{\code{Var2}}{a numeric vector}
    \item{\code{Var3}}{a numeric vector}
    \item{\code{Var4}}{a numeric vector}
    \item{\code{Var5}}{a numeric vector}
    \item{\code{Var6}}{a numeric vector}
    \item{\code{Var7}}{a numeric vector}
  }
}
\details{
  
}
\source{
  
}
\references{
  
}
\examples{
data(Future1)
## maybe str(Future1), plot(Future1) 
}
\keyword{datasets}
