\name{KappaSRE}
\alias{KappaSRE}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}

\usage{
KappaSRE(Obs, Fit, TSS = FALSE)
}

\arguments{
  \item{Obs}{  }
  \item{Fit}{  }
  \item{TSS}{  }
}
\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }