\name{KappaStat}
\alias{KappaStat}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
KappaStat(Misc)
}

\arguments{
  \item{Misc}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
