\name{Projection.raster}
\alias{Projection.raster}

\title{ Making projection with BIOMOD using raster datasets }
\description{
  For all the models currently implemented, BIOMOD is able to project potential distributions of
 species or land-use classes for other areas, other resolutions or other time scales on rasters.
}
\usage{
Projection.raster(RasterProj=NULL, Proj.name, GLM=TRUE, GBM=TRUE, GAM=TRUE, CTA=TRUE, ANN=TRUE, SRE=TRUE, quants=0.05, FDA=TRUE, MARS=TRUE, RF=TRUE, BinRoc=FALSE, BinKappa=FALSE, BinTSS=FALSE, FiltRoc=FALSE, FiltKappa=FALSE, FiltTSS=FALSE, repetition.models=TRUE, stack.out=TRUE}

\arguments{
  \item{RasterProj}{ an object of class 'RasterStack' containing the same variables than for calibrating the models (using the Models function) (names are case-sensitive !) 
   onto which to apply the models for projection making (another geographical space, resolution, or time scale). }
  \item{Proj.name}{ The name of your projection dataset (or any name you want to give to it) which will be used for storing the outputs of the function and
  also used by other functions to automatically restore the outputs in R from the hard disk}
  \item{GLM}{ if TRUE, the model will be used to render projections }
  \item{GAM}{ if TRUE, the model will be used to render projections }
  \item{GBM}{ if TRUE, the model will be used to render projections }
  \item{CTA}{ if TRUE, the model will be used to render projections }
  \item{ANN}{ if TRUE, the model will be used to render projections }
  \item{SRE}{ if TRUE, the model will be used to render projections }
  \item{quants}{ the value defines the extreme percentiles of the response data not to be used by the SRE for calibration (see \code{\link{SRE}} for more details) }
  \item{FDA}{ if TRUE, the model will be used to render projections }
  \item{MARS}{ if TRUE, the model will be used to render projections }
  \item{RF}{ if TRUE, the model will be used to render projections } 
  \item{BinRoc}{ set it to TRUE if you want the results to be also produced in binary using the Roc curve threshold for all models }
  \item{BinKappa}{ set it to TRUE if you want the results to be also produced in binary using Kappa threshold for all models }
  \item{BinTSS}{ set it to TRUE if you want the results to be also produced in binary using TSS threshold for all models }
  \item{FiltRoc}{ set it to TRUE if you want the results to be also produced as filtered values using the Roc curve threshold for all models }
  \item{FiltKappa}{ set it to TRUE if you want the results to be also produced as filtered value using Kappa threshold for all models }
  \item{FiltTSS}{ set it to TRUE if you want the results to be also produced as filtered value using TSS threshold for all models }
  \item{repetition.models}{ type TRUE to render projections on the 100 percent calibrated model and also on the repetition models. }
  \item{stack.out}{ if TRUE, the resulting projection rasters will be stored in stacks, one per model. If FALSE, every single raster projection
   will be saved individually. }
}

\Warning{
  From one R session to another and as a general reminder, do think about reloading the objects stored in the .RData file produced by the Models() function.
  They contain vital information for the models to work properly when called by other functions such as Projection().
}

\details{
  Note that the models set to TRUE need to have been run in the Models function to be able to run with Projection(). 
  Else, they simply won't be taken into account.
  
  Also take in consideration that the models can only render predictions and/or projections using the same set of variables 
  than the ones used to calibrate them. For the variable names, mind that R is case-sensitive.
}

\value{
  No values are returned but a series of objects are produced in the appropriate directory automatically generated by the function itself.
  The directory name is of the type "proj.Proj.name". If you run several projections campaign with different datasets, a directory will hence
  be produced per projection that is being run. Note that if Projection.raster() is run with the same Proj.name argument, the results will overwright.
  
  The files, single raster layers or raster stacks, are produced per species and contain the projections in binary or filtered format for all the models 
  selected and for the repetitions if wanted.
}

\references{ }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{ }
\seealso{ \code{\link{Models}}, \code{\link{ProjectionBestModel}}, \code{\link{Ensemble.Forecasting}}, \code{\link{ProbDensFunc}} }
\examples{

}
\keyword{ models }
\keyword{ regression }
\keyword{ nonlinear }
\keyword{ multivariate }
\keyword{ nonparametric }
\keyword{ tree }

