\name{Rescaler2}
\alias{Rescaler2}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}

\usage{
Rescaler2(DataToRescale, type = "range", OriMinMax = NULL)
}

\arguments{
  \item{DataToRescale}{ }
  \item{type}{ }
  \item{OriMinMax}{ }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
