\name{Rescaler4}
\alias{Rescaler4}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes.
Rescaling function using a GLM model for logistic transformation.
}

\usage{
Rescaler4(dataToRescale, ref=NULL, run, original=FALSE)
}

\arguments{
  \item{dataToRescale}{ }
  \item{ref}{ }
  \item{run}{ }
  \item{original}{ }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
