\name{SampleMat2}
\alias{SampleMat2}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
SampleMat2(ref, ratio)
}

\arguments{
  \item{ref}{  }
  \item{ratio}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }