\name{scope}
\alias{scope}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
scope(enviroTrain, Smoother, degree)
}

\arguments{
  \item{enviroTrain}{ }
  \item{Smoother}{  }
  \item{degree}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }

