\name{scopeExpSyst}
\alias{scopeExpSyst}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
scopeExpSyst(enviroTrain, mod)
}
\arguments{
  \item{enviroTrain}{  }
  \item{mod}{  }
}
\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
