\name{testnull}
\alias{testnull}

\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}
\usage{
testnull(object, Prev, dat)
}

\arguments{
  \item{object}{  }
  \item{Prev}{  }
  \item{dat}{ }
}


\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }
