\name{BIOMOD-package}
\alias{BIOMOD-package}
\alias{BIOMOD}
\docType{package}
\title{ Species/climate modelling functions 
}
\description{
BIOMOD offers the possibility to run 9 state-of-the-art models for describing
the relationship between a given species and its environment. It is an attempt
to define the ecological niche of a particular species using environmental 
variables (temperature, precipitation, ...) with the potential use of making, 
for instance, future projections under climate and land use change scenarios.
Alghouth is has been developed for ecologists to predict species distribution, 
BIOMOD can also be used to model any binomial data (for instance, gene, molecule,
ecosystem...) in function of any explanatory variables. 
}
\details{
\tabular{ll}{
Package: \tab BIOMOD\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2009-01-15\cr
License: \tab GPL \cr
}

}
\author{
Wilfried Thuiller, with participation of Bruno Lafourcade and Robin Engler
<wilfried.thuiller@ujf-grenoble.fr>
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{
 
}
