\name{Biomod.Manual}
\alias{Biomod.Manual}

\title{ Calling the Manual }
\description{ Directly opens the pdf manual fom the R session
}

\usage{
Biomod.Manual(manual="Biomod Manual"))
}

\arguments{
\item{manual}{ The name (minus the .pdf extension) of the file you want to open }
}


\author{ Wilfried Thuiller, Bruno Lafourcade }

\seealso{ \code{\link{help}}}
\examples{

Biomod.Manual()
Biomod.Manual("Biomod_Presentation_Manual")
Biomod.Manual("Biomod_Practical_Building_Models")
}

\keyword{ models }
