\name{Biomod.Manual}
\alias{Biomod.Manual}

\title{ Calling the Manual }
\description{ Directly opens the pdf manual fom the R session
}

\usage{
Biomod.Manual()
}

\arguments{
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{

Biomod.Manual()
}

\keyword{ models }
