\name{Future1}
\alias{Future1}
\docType{data}
\title{ Future data for making projections }
\description{ A simulated climate change scenario for the same area as Sp.Env. It contains the same
7 climate variables (necessary to run the projection functions) and for the same resolution (50x50km)
  
}
\usage{data(Future1)}
\format{
  A data frame with 2264 observations on the following 10 variables.
  \describe{
    \item{\code{Idw}}{an Id of the sites to be matched with the one of Sp.Env}
    \item{\code{X}}{longitude of the sites}
    \item{\code{Y}}{latitude of the sites} 
    \item{\code{Var1}}{a climate variable}
    \item{\code{Var2}}{a climate variable}
    \item{\code{Var3}}{a climate variable}
    \item{\code{Var4}}{a climate variable}
    \item{\code{Var5}}{a climate variable}
    \item{\code{Var6}}{a climate variable}
    \item{\code{Var7}}{a climate variable}
  }
}
\details{
  
}
\source{
  
}
\references{
  
}
\examples{
data(Future1)
fix(Future1)
 
}
\keyword{datasets}
