\name{KappaRepet}
\alias{KappaRepet}


\title{ Inner Function }
\description{ Inner function for the BIOMOD package, not for independent usage purposes
}

\usage{
KappaRepet(Obs, Fit, TSS = FALSE)
}

\arguments{
  \item{Obs}{  }
  \item{Fit}{  }
  \item{TSS}{  }
}

\details{
  
}
\value{
  
}
\references{  }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{  


}
\seealso{ \code{\link{help}}}
\examples{
}

\keyword{ models }