\name{KappaStat}
\alias{KappaStat}

\title{  Kappa Statistic for predictive accuracy evaluation }
\description{ Calculate the True Skill Statistic to evaluate the predictive accuracy of a given species distribution model
}
\usage{
KappaStat(Misc)
}

\arguments{
  \item{Misc}{ A misclassification table, counting the number of true and false positive and true and false negative  }
}


\author{ Wilfried Thuiller, Bruno Lafourcade }

\seealso{ \code{\link{help}}}

\keyword{ models }
