\name{Projection}
\alias{Projection}

\title{ Making projection with BIOMOD }
\description{
  For all the models currently implemented, BIOMOD is able to project potential distributions of
species or land-use classes for other areas, other resolutions or other time scales.
}
\usage{
Projection(Proj = NULL, Proj.name, GLM = TRUE, GBM = TRUE, GAM = TRUE, CTA = TRUE, ANN = TRUE, SRE = TRUE, Perc025 = FALSE, Perc05 = TRUE, MDA = TRUE, MARS = TRUE, RF = TRUE, BinRoc = FALSE, BinKappa = FALSE, BinTSS = FALSE, FiltRoc = FALSE, FiltKappa = FALSE, FiltTSS = FALSE, repetition.models=TRUE)
}

\arguments{
  \item{Proj}{ a matrix containing the same variables than for calibrating the models (using the Models function) (the column names are case-sensitive !) 
   but on another geographical space, resolution, or time scale. }
  \item{Proj.name}{ The name of your projection dataset (or any name you want to give to it) which will be used for storing the outputs of the function and
  also used by other functions to automatically restore the outputs in R from the hard disk}
  \item{GLM}{ if TRUE, the model will be used to render projections }
  \item{GAM}{ if TRUE, the model will be used to render projections }
  \item{GBM}{ if TRUE, the model will be used to render projections }
  \item{CTA}{ if TRUE, the model will be used to render projections }
  \item{ANN}{ if TRUE, the model will be used to render projections }
  \item{SRE}{ if TRUE, the model will be used to render projections }
  \item{Perc025}{ if TRUE, the 0.025 extreme percentiles of the response data will not be used by the SRE for calibration}
  \item{Perc05}{ if TRUE, the 0.05 extreme percentiles of the response data will not be used by the SRE for calibration}
  \item{MDA}{ if TRUE, the model will be used to render projections }
  \item{MARS}{ if TRUE, the model will be used to render projections }
  \item{RF}{ if TRUE, the model will be used to render projections } 
  \item{BinRoc}{ set it to TRUE if you want the results to be also produced in binary using the Roc curve threshold for all models }
  \item{BinKappa}{ set it to TRUE if you want the results to be also produced in binary using Kappa threshold for all models }
  \item{BinTSS}{ set it to TRUE if you want the results to be also produced in binary using TSS threshold for all models }
  \item{FiltRoc}{ set it to TRUE if you want the results to be also produced as filtered values using the Roc curve threshold for all models }
  \item{FiltKappa}{ set it to TRUE if you want the results to be also produced as filtered value using Kappa threshold for all models }
  \item{FiltTSS}{ set it to TRUE if you want the results to be also produced as filtered value using TSS threshold for all models }
  \item{repetition.models}{ type TRUE to render projections on the 100 percent calibrated model and also on the repetition models. }
}

\Warning{
  From one R session to another and as a general reminder, do think about reloading the objects stored in the .RData file produced by the Models() function.
  They contain vital information for the models to work properly when called by other functions such as Projection().
}

\details{
  Note that the models set to TRUE need to have been run in the Models function to be able to run with Projection(). 
  Else, they simply won't be taken into account.
  
  Also take in consideration that the models can only render predictions and/or projections using the same set of variables 
  than the ones used to calibrate them. For the variable names, mind that R is case-sensitive.
}

\value{
  No values are returned but a series of objects are produced in the appropriate directory automatically generated by the function itself.
  The directory name is of the type "proj.Proj.name". If you run several projections campaign with different datasets, a directory will hence
  be produced per projection that is being run. Note that if Projection() is run with the same Proj.name argument, the results will overwright.
  
  The files (arrays with 4 dimensions) are produced per species and contain the projections in binary or filtered format for all the models 
  selected and for the repetitions if wanted. See the structure in the examples below.
}

\references{ }
\author{ Wilfried Thuiller, Bruno Lafourcade }
\note{ }
\seealso{ \code{\link{Models}}, \code{\link{ProjectionBestModel}}, \code{\link{Ensemble.Forecasting}}, \code{\link{ProbDensFunc}}, \code{\link{map.plot}} }
\examples{

data(Sp.Env)
data(CoorXY)

#This command is necessary for the run of BIOMOD as a new dataframe is produced for the Models function
Initial.State(Response=Sp.Env[,13:14], Explanatory=Sp.Env[,4:10], 
IndependentResponse=NULL, IndependentExplanatory=NULL)

#Here are done 2 PA runs and 2 repetitions for each run. This will hence take several minutes.
Models(GLM = TRUE, TypeGLM = "quad", Test = "AIC", CTA = TRUE, CV.tree = 50, ANN = TRUE, CV.ann = 2, 
   NbRunEval = 2, DataSplit = 80, Roc=TRUE, Optimized.Threshold.Roc=TRUE, Kappa=TRUE, TSS=TRUE, VarImport=5,
   NbRepPA=0, strategy="circles", coor=CoorXY, distance=2, nb.absences=1000)


#Render Future projections under climate change scenario
data(Future1)
head(Future1) #the head() function enables to visualise the 5 firts lines of a dataframe
Projection(Proj = Future1[,4:10], Proj.name="Future1", GLM=TRUE, CTA=TRUE, ANN=TRUE, BinKappa=TRUE, FiltKappa=TRUE)


#view the results produced, example with the projections in probabilities :
load("proj.Future1/Proj_Future1_Sp277")
Proj_Future1_Sp277[100:150,,1,1]


#and plot them
data(CoorXY)
load("proj.Future1/Proj_Future1_Sp277_BinKappa")
load("proj.Future1/Proj_Future1_Sp277_FiltKappa")

par(mfrow=c(1,4))
level.plot(Sp.Env[,13], CoorXY, show.scale=FALSE, title="original data")
level.plot(Proj_Future1_Sp277[,"GLM",2,1], CoorXY, show.scale=FALSE, title="probabilities by GLM")
level.plot(Proj_Future1_Sp277_BinKappa[,"GLM",2,1], CoorXY, show.scale=FALSE, title="binary by Kappa")
level.plot(Proj_Future1_Sp277_FiltKappa[,"GLM",2,1], CoorXY, show.scale=FALSE, title="filtered by Kappa")


}
\keyword{ models }
\keyword{ regression }
\keyword{ nonlinear }
\keyword{ multivariate }
\keyword{ nonparametric }
\keyword{ tree }

