\name{TSS.Stat}
\alias{TSS.Stat}

\title{ True Skill Statistic for predictive accuracy evaluation }
\description{ Calculate the True Skill Statistic to evaluate the predictive accuracy of a given species distribution model
}

\usage{
TSS.Stat(Misc)
}

\arguments{
  \item{Misc}{A misclassification table, counting the number of true and false positive and true and false negative }
}



\references{Allouche O., Tsoar A. & Kadmon R. (2006). Assessing the accuracy of species distribution models: prevalence, kappa and the true skill statistic (TSS). Journal of Applied Ecology, 43, 1223-1232.
  }
\author{ Wilfried Thuiller, Bruno Lafourcade }



\keyword{ models }
